% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arslan}
\alias{arslan}
\alias{arslan_3d}
\title{arslan atlas}
\format{
An object of class \code{brain_atlas} of length 4.

An object of class \code{ggseg3d_atlas} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4 rows and 4 columns.
}
\usage{
arslan

arslan_3d
}
\description{
Add a short description of this atlas.
}
\examples{
data(arslan)
data(arslan_3d)
}
\references{
Arslan, S., & Rueckert, D. (2015, October). 
    Multi-level parcellation of the cerebral cortex using resting-state fMRI. 
    In International Conference on Medical Image Computing and Computer-Assisted 
    Intervention (pp. 47-54). Springer, Cham.
    (\href{https://link.springer.com/chapter/10.1007/978-3-319-24574-4_6}{Springer})
    
\itemize{
 \item{arslan}{ - arslan atlas}
 \item{arslan_3d}{ - arslan 3d mesh atlas}
}
}
\concept{ggseg_atlases ggseg3d_atlases}
\keyword{datasets}
