% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arslan}
\alias{arslan}
\title{Arslan Multi-Level Parcellation Atlas}
\format{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\usage{
data(arslan)
}
\description{
Brain atlas for the Arslan multi-level parcellation of the cerebral cortex
using resting-state fMRI with 50 regions. Contains both 2D polygon
geometry for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}} and 3D vertex indices for
\code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}.
}
\examples{
data(arslan)
arslan
}
\references{
Arslan, S., & Rueckert, D. (2015, October).
Multi-level parcellation of the cerebral cortex using resting-state fMRI.
In International Conference on Medical Image Computing and
Computer-Assisted Intervention (pp. 47-54). Springer, Cham.
\doi{10.1007/978-3-319-24574-4_6}
}
\concept{ggseg_atlases}
\keyword{datasets}
